backend/admin/requirements.txt
asgiref==3.8.1
boto3==1.34.132
botocore==1.34.132
Django==5.0.6
django-cors-headers==4.4.0
django-storages==1.14.3
djangorestframework==3.15.2
jmespath==1.0.1
mysqlclient==2.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
